Normalization of misnamed properties. Addresses issue with annotations being improperly generated. Addresses missing module name issues in tests.

+++ NEsper/NEsper.Regression/suite/expr/exprcore/ExprCoreRelOp.cs
@@ -74,7 +74,7 @@ namespace com.espertech.esper.regressionlib.suite.expr.exprcore
                 bean => bean.IntPrimitive = 3);
             RunAssertion(
                 env,
-                "bigInteger",
+                "BigInteger",
                 "BigInteger.ValueOf(2)",
                 bean => bean.BigInteger = new BigInteger(1),
                 bean => bean.BigInteger = new BigInteger(2),
